make `Controls` UI component instead of `Controls` colors block
