+++ .changeset/hot-mayflies-prove.md
@@ -1,0 +1,5 @@
+---
+'snable': patch
+---
+
+make `Controls` UI component instead of `Controls` colors block
